feat(*)!: Introduces a new wadm-types crate
This allows us to import the types separately from the rest of the lib
code for wadm
